change filter reg back

+++ include/4131.h
@@ -57,11 +57,11 @@ void write_mem_map(void);
 void spi_read_reg(uint8_t* rx_data, uint8_t reg_addr, size_t bytes);
 void read_adc_id(void);
 void read_adc_conversion(void);
-uint32_t get_adc_data(void);
+uint16_t get_adc_data(void);
 uint8_t read_status(void);
-uint32_t bytes_to_dec(uint8_t *bytes);
+uint16_t bytes_to_dec(uint8_t *bytes);
 void test_spi_send(void);
-uint32_t hex_to_code(const uint8_t *buff, size_t len);
-uint32_t get_data_from_channel(uint8_t channel_idx);
+uint16_t hex_to_code(const uint8_t *buff, size_t len);
+uint16_t get_data_from_channel(uint8_t channel_idx);
 void configure_adc_channel(uint8_t channel_idx, uint8_t enable_bit);
 #endif // 4131_H_
